unified all components in one folder
