+++ app/carrito/page.tsx
@@ -1,11 +1,9 @@
 "use client";
-
 import { useState } from "react";
-import Link from "next/link";
 import Image from "next/image";
-import { TrashIcon, PlusIcon } from "@heroicons/react/24/outline";
-import NavbarWhite from "../components/navbarWhite";
-import Footer from "../components/footer";
+import { TrashIcon } from "@heroicons/react/24/outline";
+import NavbarWhite from "@/components/navbarWhite";
+import Footer from "@/components/footer";
 
 interface CartItem {
   id: number;
