+++ app/pedidos/page.tsx
@@ -21,7 +21,7 @@ import {
   CheckSquare,
   Square,
 } from "lucide-react";
-import Sidebar from "../components/admins/sidebar";
+import Sidebar from "@/components/sidebar";
 
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
@@ -46,7 +46,6 @@ import {
   Dialog,
   DialogContent,
   DialogDescription,
-  DialogFooter,
   DialogHeader,
   DialogTitle,
   DialogTrigger,
@@ -63,7 +62,7 @@ import {
   AlertDialogTitle,
   AlertDialogTrigger,
 } from "@/components/ui/alert-dialog";
-import OrderFormModal from "../components/orderFormModal";
+import OrderFormModal from "@/components/orderFormModal";
 
 interface Product {
   producto_id: number;
@@ -536,7 +535,10 @@ const OrderDashboard: React.FC = () => {
                             >
                               {selectedOrders.length === currentOrders.length &&
                               currentOrders.length > 0 ? (
-                                <CheckSquare size={18} className="text-black" />
+                                <CheckSquare
+                                  size={18}
+                                  className="text-blue-600"
+                                />
                               ) : (
                                 <Square size={18} className="text-gray-400" />
                               )}
@@ -608,7 +610,10 @@ const OrderDashboard: React.FC = () => {
                               className="focus:outline-none"
                             >
                               {selectedOrders.includes(order.pedido_id) ? (
-                                <CheckSquare size={18} className="text-black" />
+                                <CheckSquare
+                                  size={18}
+                                  className="text-blue-600"
+                                />
                               ) : (
                                 <Square size={18} className="text-gray-400" />
                               )}
@@ -792,7 +797,10 @@ const OrderDashboard: React.FC = () => {
                               className="mr-2 focus:outline-none"
                             >
                               {selectedOrders.includes(order.pedido_id) ? (
-                                <CheckSquare size={18} className="text-black" />
+                                <CheckSquare
+                                  size={18}
+                                  className="text-blue-600"
+                                />
                               ) : (
                                 <Square size={18} className="text-gray-400" />
                               )}
